Add files via upload
What is Stack Overflow in Recursion?

Whenever recursion calls are executed, they’re simultaneously stored in a recursion stack where they wait to complete the recursive function. A recursive function can only be completed if a base condition is fulfilled and the control returns to the parent function. 

But, when there is no base condition given for a particular recursive function, it gets called indefinitely which results in a Stack Overflow i.e, exceeding the memory limit of the recursion stack. Hence, the program terminates giving a Segmentation Fault error. 

the basics of recursion such as the following were clear to us : 

1. What is Recursion
2. Base Condition
3. Stack Overflow / Stack Space
4. Recursion Tree